refactor(run download): extract zip extraction func into a separate package
